💄 Refactor: smooth scroll 전체 페이지에 적용
smooth scroll 모든 페이지에 적용
