+++ .vscode/settings.json
@@ -52,6 +52,7 @@
 		"SSERAFIM",
 		"SULLYOON",
 		"Testpage",
+		"Toastify",
 		"uuidv",
 		"WONYOUNG",
 		"YUJIN",
